feat(ui): add ImportDialog component for Bazarr settings
Implement a new dialog for importing settings from Bazarr,
enhancing user experience for configuration management.

Files changed:
- Added ImportDialog component for Bazarr settings import: [src/components/ImportDialog.jsx](https://github.com/jdfalk/subtitle-manager/blob/main/webui/src/components/ImportDialog.jsx)
- Updated Settings component to utilize ImportDialog: [src/Settings.jsx](https://github.com/jdfalk/subtitle-manager/blob/main/webui/src/Settings.jsx)
- Updated README with common issues for rebase scripts: [scripts/README-rebase.md](https://github.com/jdfalk/subtitle-manager/blob/main/scripts/README-rebase.md)
